+++ src/mofaflex/_core/datasets/__init__.py
@@ -1,5 +1,5 @@
 from .anndatadataset import AnnDataDataset
 from .anndatadictdataset import AnnDataDictDataset
 from .base import MofaFlexDataset, Preprocessor
-from .misc import CovariatesDataset, MofaFlexBatchSampler, StackDataset, merge_covariates
+from .misc import CovariatesDataset, MofaFlexBatchSampler, StackDataset, df_to_array, merge_covariates
 from .mudatadataset import MuDataAxis0Dataset, MuDataAxis1Dataset